*: remove model v1 (#746)
