+++ feeluown/gui/components/avatar.py
@@ -4,7 +4,7 @@ from PyQt5.QtWidgets import QMenu, QAction
 from PyQt5.QtGui import QPainter, QIcon, QPalette, QContextMenuEvent
 
 from feeluown.library import NoUserLoggedIn, UserModel
-from feeluown.models.uri import reverse
+from feeluown.library import reverse
 from feeluown.utils.aio import run_afn, run_fn
 from feeluown.gui.provider_ui import UISupportsLoginOrGoHome, ProviderUiItem, \
     UISupportsLoginEvent
