+++ feeluown/gui/mimedata.py
@@ -1,6 +1,6 @@
 from PyQt5.QtCore import QMimeData
 
-from feeluown.models import ModelType
+from feeluown.library import ModelType
 
 
 model_mimetype_map = {
