+++ feeluown/player/playlist.py
@@ -15,7 +15,7 @@ from feeluown.library import (
     MediaNotFound, SongProtocol, ModelType, NotSupported, ResourceNotFound
 )
 from feeluown.media import Media
-from feeluown.models.uri import reverse
+from feeluown.library import reverse
 
 if TYPE_CHECKING:
     from feeluown.app import App
@@ -81,7 +81,7 @@ class Playlist:
     def __init__(self, app: 'App', songs=None, playback_mode=PlaybackMode.loop,
                  audio_select_policy='hq<>'):
         """
-        :param songs: list of :class:`feeluown.models.SongModel`
+        :param songs: list of :class:`feeluown.library.SongModel`
         :param playback_mode: :class:`feeluown.player.PlaybackMode`
         """
         self._app = app
