+++ tests/library/test_library.py
@@ -2,15 +2,7 @@ import random
 
 import pytest
 
-from feeluown.library import Library, ModelType, BriefAlbumModel
-from feeluown.library.provider import dummy_provider
-from feeluown.models import SearchModel
-
-
-def test_library_search(library):
-    result = list(library.search('xxx'))[0]
-    assert len(result.songs) >= 3
-    assert result.songs[0].identifier == 1
+from feeluown.library import ModelType, BriefAlbumModel
 
 
 @pytest.mark.asyncio
@@ -63,60 +55,6 @@ def test_score_fn():
         create_song('很爱很爱你', '刘若英', '脱掉高跟鞋世界巡回演唱会', '05:24')
     ]
     assert default_score_fn(song, candidates[0]) >= MIN_SCORE
-
-
-def test_library_list_songs_standby(library, song):
-    songs = library.list_song_standby(song)
-
-    # all songs share the same provider,
-    # so there will be no standby song
-    assert len(songs) == 0
-
-    song.source = 'dummy-1'
-    songs = library.list_song_standby(song)
-    assert len(songs) == 1
-
-    songs = library.list_song_standby(song, onlyone=False)
-    assert len(songs) == 2
-
-
-@pytest.mark.asyncio
-async def test_library_a_list_songs_standby(library, song):
-    songs = await library.a_list_song_standby(song)
-    assert len(songs) <= 1
-
-    song.source = 'dummy-1'
-    songs = await library.a_list_song_standby(song)
-    assert len(songs) == 1
-
-
-@pytest.mark.asyncio
-async def test_library_a_list_songs_standby_with_specified_providers(song):
-    library = Library(providers_standby=['xxx'])
-    song.source = 'dummy-1'
-    songs = await library.a_list_song_standby(song)
-    assert len(songs) == 0
-
-
-@pytest.mark.asyncio
-async def test_library_a_list_songs_standby_v2(library, provider,
-                                               song, song1, song_standby, mocker):
-    mock_search = mocker.patch.object(provider, 'search')
-    mock_search.return_value = SearchModel(q='xx', songs=[song1, song_standby])
-
-    song_media_list = await library.a_list_song_standby_v2(song)
-    for standby, media in song_media_list:
-        if standby is song_standby:
-            assert media.url == 'standby.mp3'
-            break
-    else:
-        assert False, 'song_standby should be a stanby option'
-
-
-def test_library_register_should_emit_signal(library, mocker):
-    mock_emit = mocker.patch('feeluown.utils.dispatch.Signal.emit')
-    library.register(dummy_provider)
-    mock_emit.assert_called_once_with(dummy_provider)
 
 
 def test_library_model_get(library, ekaf_provider, ekaf_album0):
