+++ tests/serializers/test_serializers.py
@@ -1,10 +1,7 @@
-import pytest
-
 from feeluown.app import App
 from feeluown.player import Player, Playlist
 from feeluown.serializers import serialize
 from feeluown.library import SongModel
-from feeluown.models import SongModel as SongModelV1
 from feeluown.player import Metadata
 
 
@@ -37,11 +34,4 @@ def test_serialize_model():
     assert song_js['identifier'] == '1'
 
     song_js = serialize('python', song, fetch=True)
-    assert song_js['identifier'] == '1'
-
-
-@pytest.mark.skip(reason='model v1 will be removed lator')
-def test_serialize_model_v1():
-    song = SongModelV1(identifier='1', title='', artists=[], duration=0)
-    song_js = serialize('python', song)
     assert song_js['identifier'] == '1'
